+++ frontend/packages/app/src/pages/dashboard/widget/heatmap/index.tsx
@@ -18,7 +18,7 @@ import { useFrappeGetCall } from "frappe-react-sdk";
  * Internal dependencies.
  */
 import { HeatmapCardSkeleton } from "./skeleton";
-import type { AllocationHeatmapResponse, RoleAllocationWeek } from "../types";
+import type { AllocationHeatmapResponse, RoleAllocationWeek } from "./types";
 
 export type HeatmapCellState = "full" | "partial" | "none";
 
@@ -46,7 +46,7 @@ function getCellState(week: RoleAllocationWeek): HeatmapCellState {
   return "partial";
 }
 
-export function HeatmapCard() {
+export default function HeatmapCard() {
   const [selectedRoles, setSelectedRoles] = useState<string[]>([]);
 
   const args = useMemo(() => {
