+++ frontend/packages/app/src/pages/dashboard/widget/heatmap/types.ts
@@ -1,0 +1,23 @@
+export interface WeekRange {
+  week_start: string;
+  week_end: string;
+}
+
+export interface RoleAllocationWeek extends WeekRange {
+  capacity_hours: number;
+  allocated_hours: number;
+}
+
+export interface RoleAllocation {
+  designation: string;
+  weeks: RoleAllocationWeek[];
+}
+
+export interface AllocationHeatmapResponse {
+  message: {
+    from_date: string;
+    to_date: string;
+    weeks: WeekRange[];
+    roles: RoleAllocation[];
+  };
+}
