+++ frontend/packages/app/src/pages/dashboard/widget/kpi-cards/skeleton.tsx
