+++ frontend/packages/app/src/pages/dashboard/widget/kpi-cards/types.ts
@@ -1,0 +1,14 @@
+export interface KPIMetric {
+  current: number;
+  previous: number;
+  change_pct: number | null;
+  trend: "up" | "down";
+}
+
+export interface LeadershipKPIResponse {
+  message: {
+    revenue: KPIMetric;
+    cost: KPIMetric;
+    profit_margin: KPIMetric;
+  };
+}
