+++ frontend/packages/app/src/pages/dashboard/widget/stat-cards/skeleton.tsx
