+++ frontend/packages/app/src/pages/dashboard/widget/stat-cards/stat-card.tsx
