+++ frontend/packages/app/src/pages/dashboard/widget/utilization/types.ts
@@ -1,0 +1,7 @@
+export interface TimeUtilisationResponse {
+  message: {
+    billable_hours: number;
+    non_billable_hours: number;
+    total_hours: number;
+  };
+}
